requirements.txt
attrs==21.2.0
colour==0.1.5
defusedxml==0.7.1
et-xmlfile==1.1.0
jsonschema==4.1.2
numpy==1.21.3
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.3.4
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
styleframe==4.0.0
xlrd==1.2.0
XlsxWriter==3.0.2
